Move XCTestCase snapshot helpers to separate file
